SellerManagement/src/main/java/com/example/sellermanagement/entity/GetUsers.java
package com.example.sellermanagement.entity;

public class GetUsers {

    private int id;
    private String userName;
    private String dob;
    private String registration;
    private String status = "PENDING";
    private String name;

    public int getId() {
        return id;
    }

    public void setId(int id) {
        this.id = id;
    }

    public String getUserName() {
        return userName;
    }

    public void setUserName(String userName) {
        this.userName = userName;
    }

    public String getDob() {
        return dob;
    }

    public void setDob(String dob) {
        this.dob = dob;
    }

    public String getRegistration() {
        return registration;
    }

    public void setRegistration(String registration) {
        this.registration = registration;
    }

    public String getStatus() {
        return status;
    }

    public void setStatus(String status) {
        this.status = status;
    }

    public String getName() {
        return name;
    }

    public void setName(String name) {
        this.name = name;
    }

    @Override
    public String toString() {
        return "GetUsers{" +
                "id=" + id +
                ", userName='" + userName + '\'' +
                ", dob='" + dob + '\'' +
                ", registration='" + registration + '\'' +
                ", status='" + status + '\'' +
                ", name='" + name + '\'' +
                '}';
    }
}
